Split Box runtime family definitions
